sync update

+++ src/redux/sync.ts
@@ -126,10 +126,14 @@ export const syncWalletAddresses = createAsyncThunk(
 
     AddressScanner.populateAddresses();
 
+    const addresses = AddressManager.getReceiveAddresses();
+    thunkApi.dispatch(syncSubscriptionCount(addresses.length));
+
     Promise.all(
-      AddressManager.getReceiveAddresses().map((address) =>
-        Electrum.subscribeToAddress(address)
-      )
+      addresses.map(async (address) => {
+        await Electrum.subscribeToAddress(address);
+        thunkApi.dispatch(syncSubscriptionCount(-1));
+      })
     );
 
     thunkApi.dispatch(syncChangeAddresses());
@@ -148,7 +152,12 @@ export const syncChangeAddresses = createAsyncThunk(
 
     const promises = changeAddresses
       .filter((address) => !(address.state !== null && address.balance === 0)) // don't resync fully spent change addresses
-      .map((address) => Electrum.subscribeToAddress(address));
+      .map(async (address) => {
+        await Electrum.subscribeToAddress(address);
+        thunkApi.dispatch(syncSubscriptionCount(-1));
+      });
+
+    thunkApi.dispatch(syncSubscriptionCount(promises.length));
 
     const batchedPromises = [];
     const batch_chunk_size = 1024;
@@ -162,6 +171,10 @@ export const syncChangeAddresses = createAsyncThunk(
       Promise.all(batch);
     });
   }
+);
+
+export const syncSubscriptionCount = createAction<number>(
+  "sync/subscriptionCount"
 );
 
 // syncAddressState: fired when data acquired from address subscription
@@ -314,51 +327,43 @@ export const syncHotRefresh = createAsyncThunk(
   }
 );
 
-export const syncBlock = createAsyncThunk(
-  "sync/block",
-  async (height: number) => {
-    const Blockchain = BlockchainService();
-    let block;
-    try {
-      block = await Blockchain.getBlockByHeight(height);
-    } catch {
-      const header = await Electrum.requestBlock(height);
-      await Blockchain.registerBlock({ header, height });
-      block = await Blockchain.getBlockByHeight(height);
+export const syncChaintip = createAsyncThunk(
+  "sync/chaintip",
+  async (chaintip: { height: number; hex: string }, thunkApi) => {
+    const sync = selectSyncState(thunkApi.getState());
+    if (sync.syncPending.chaintip > 1) {
+      return sync.chaintip;
     }
 
-    Log.log("sync/block", block);
+    const Blockchain = BlockchainService();
+    const block = await Blockchain.resolveBlockByHeight(chaintip.height);
+
+    Log.log("sync/chaintip", block);
+
+    queueMicrotask(async () => {
+      await TransactionManagerService().purgeTransactions();
+      await BlockchainService().purgeBlocks();
+    });
+
     return block;
   }
 );
 
-export const syncChaintip = createAction<{ height: number; hex: string }>(
-  "sync/chaintip"
-);
-syncMiddleware.startListening({
-  actionCreator: syncChaintip,
-  effect: async (action, listenerApi) => {
-    const chaintip = action.payload;
-    Log.log("sync/chaintip", chaintip);
-    listenerApi.dispatch(syncBlock(chaintip.height));
-
-    await TransactionManagerService().purgeTransactions();
-    BlockchainService().purgeBlocks();
-  },
-});
-
 const initialPending = {
   utxo: 0,
   history: 0,
-  txState: 0,
-  change: 0,
+  addressState: 0,
+  subscription: 0,
+  chaintip: 0,
   rebuild: 0,
 };
 const initialState = {
   isConnected: false,
   server: "",
   syncPending: { ...initialPending },
-  chaintip: { ...block_checkpoints.first2023 },
+  chaintip: {
+    ...block_checkpoints.first2023,
+  },
   addresses: {},
   lastRefresh: Date.now(),
 };
@@ -376,15 +381,15 @@ export const syncReducer = createReducer(initialState, (builder) => {
       state.isConnected = false;
     })
     .addCase(syncAddressState.pending, (state: RootState) => {
-      state.syncPending.txState += 1;
+      state.syncPending.addressState += 1;
     })
     .addCase(syncAddressState.fulfilled, (state: RootState, action) => {
       const [address] = action.payload;
       state.addresses[address.address] = address;
-      state.syncPending.txState -= 1;
+      state.syncPending.addressState -= 1;
     })
     .addCase(syncAddressState.rejected, (state: RootState) => {
-      state.syncPending.txState -= 1;
+      state.syncPending.addressState -= 1;
     })
     .addCase(syncAddressUtxos.pending, (state: RootState) => {
       state.syncPending.utxo += 1;
@@ -413,22 +418,15 @@ export const syncReducer = createReducer(initialState, (builder) => {
     .addCase(txHistoryFetch.rejected, (state: RootState) => {
       state.syncPending.history -= 1;
     })
-    .addCase(syncChangeAddresses.pending, (state: RootState) => {
-      state.syncPending.change += 1;
+    .addCase(syncSubscriptionCount, (state, action) => {
+      state.syncPending.subscription += action.payload;
     })
-    .addCase(syncChangeAddresses.fulfilled, (state: RootState) => {
-      state.syncPending.change -= 1;
+    .addCase(syncChaintip.pending, (state) => {
+      state.syncPending.chaintip += 1;
     })
-    .addCase(syncChangeAddresses.rejected, (state: RootState) => {
-      state.syncPending.change -= 1;
-    })
-    .addCase(syncChaintip, (state, action) => {
-      const { hex, height } = action.payload;
-      state.chaintip = {
-        blockhash: BlockchainService().calculateBlockhash(hex),
-        header: hex,
-        height,
-      };
+    .addCase(syncChaintip.fulfilled, (state, action) => {
+      state.syncPending.chaintip -= 1;
+      state.chaintip = action.payload;
     })
     .addCase(syncHotRefresh.fulfilled, (state: RootState, action) => {
       state.lastRefresh = action.payload;
@@ -453,6 +451,11 @@ export const selectSyncState = createSelector(
 export const selectMyAddresses = createSelector(
   (state: RootState) => state.sync,
   (sync) => sync.addresses
+);
+
+export const selectChaintip = createSelector(
+  (state) => state.sync,
+  (sync) => sync.chaintip
 );
 
 /*
